+++ packages/desktop-core/src/main/storage/provisioners/index.ts
@@ -1,0 +1,2 @@
+export * from "../iStorageClient";
+export * from "../iStorageProvisioner";
